Refactor code and fix issue with edit mode environment object
