+++ TimeFlare/Models/Dashboard_Models/AppLaunchModel.swift
