+++ TimeFlare/Views/Onboarding/WelcomeView.swift
@@ -7,34 +7,6 @@
 
 import SwiftUI
 
-enum OnboardingItem {
-    case addDeadline
-    case selectImage
-    case widget
-    
-    var image: Image {
-        switch self {
-        case .addDeadline:
-            return .onboardingAddDeadline
-        case .selectImage:
-            return .onboardingAddImage
-        case .widget:
-            return .onboardingWidget
-        }
-    }
-    
-    var description: String {
-        switch self {
-        case .addDeadline:
-            return "Create deadlines by clicking on the plus button in the navigation bar."
-        case .selectImage:
-            return "Select an image to make each deadline more unique."
-        case .widget:
-            return "Take a glance at any ongoing deadline by adding the TimeFlare widget to your home screen."
-        }
-    }
-}
-
 struct WelcomeView: View {
     
     var mainAction: Thunk
